+++ modules/__tests__/HashHistory-test.js
@@ -247,5 +247,25 @@ describeHistory('a hash history', () => {
       const history = createHistory({ basename: '/prefix' })
       expect(history.location.pathname).toEqual('/')
     })
+
+    describe('path params', () => {
+      it('strips the basename', () => {
+        window.history.replaceState(null, null, '#/PREFIX/PARAM/pathname')
+        const history = createHistory({ basename: '/prefix/:param/pathname' })
+        expect(history.location.pathname).toEqual('/')
+      })
+
+      it('strips when path is only the prefix', () => {
+        window.history.replaceState(null, null, '#/PREFIX/PARAM/pathname')
+        const history = createHistory({ basename: '/prefix/:param' })
+        expect(history.location.pathname).toEqual('/pathname')
+      })
+
+      it('strips when path is only the prefix with multiple params', () => {
+        window.history.replaceState(null, null, '#/PREFIX/PARAM/midfix/PARAM/pathname')
+        const history = createHistory({ basename: '/prefix/:param/midfix/:param' })
+        expect(history.location.pathname).toEqual('/pathname')
+      })
+    })
   })
 })